+++ torchmetrics/image/__init__.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from torchmetrics.image.psnr import PeakSignalNoiseRatio  # noqa: F401
+from torchmetrics.image.rase import RelativeAverageSpectralError  # noqa: F401
+from torchmetrics.image.rmse_sw import RootMeanSquaredErrorUsingSlidingWindow  # noqa: F401
 from torchmetrics.image.ssim import (  # noqa: F401
     MultiScaleStructuralSimilarityIndexMeasure,
     StructuralSimilarityIndexMeasure,