+++ CHANGELOG.md
@@ -13,8 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added support for `MetricCollection` in `MetricTracker` ([#718](https://github.com/PyTorchLightning/metrics/pull/718))
 
-
-- Added new image metric `UniversalImageQualityIndex` ([#824](https://github.com/PyTorchLightning/metrics/pull/824))
+- Added new image metrics:
+  * `UniversalImageQualityIndex` ([#824](https://github.com/PyTorchLightning/metrics/pull/824))
+  * `RootMeanSquaredErrorUsingSlidingWindow`  ([#816](https://github.com/PyTorchLightning/metrics/pull/816))
+  * `RelativeAverageSpectralError` ([#816](https://github.com/PyTorchLightning/metrics/pull/816))
 
 
 - Added smart update of `MetricCollection` ([#709](https://github.com/PyTorchLightning/metrics/pull/709))